Backend projects/requirements.txt
# GitHub API Proxy Server Dependencies

flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
cryptography>=41.0.0
